+++ tests/testthat/test-joins.R
@@ -140,53 +140,104 @@ test_that("'other' must be a LazyFrame", {
 })
 
 test_that("argument 'validate' works", {
-  df1 = pl$LazyFrame(x = letters[1:5], y = 1:5)
-  df2 = pl$LazyFrame(x = c("a", letters[1:4]), y2 = 6:10)
+  df1 = pl$DataFrame(x = letters[1:5], y = 1:5)
+  df2 = pl$DataFrame(x = c("a", letters[1:4]), y2 = 6:10)
 
-  # 1:1
+  # eager 1:1
   expect_error(
-    df1$join(df2, on = "x", validate = "1:1")$collect(),
+    df1$join(df2, on = "x", validate = "1:1"),
     "join keys did not fulfil 1:1 validation"
   )
 
-  # m:1
+  # lazy 1:1
   expect_error(
-    df1$join(df2, on = "x", validate = "m:1")$collect(),
+    df1$lazy()$join(df2$lazy(), on = "x", validate = "1:1")$collect(),
+    "join keys did not fulfil 1:1 validation"
+  )
+
+  # eager m:1
+  expect_error(
+    df1$join(df2, on = "x", validate = "m:1"),
     "join keys did not fulfil m:1 validation"
   )
 
-  # 1:m
+  # lazy m:1
   expect_error(
-    df2$join(df1, on = "x", validate = "1:m")$collect(),
+    df1$lazy()$join(df2$lazy(), on = "x", validate = "m:1")$collect(),
+    "join keys did not fulfil m:1 validation"
+  )
+
+  # eager 1:m
+  expect_error(
+    df2$join(df1, on = "x", validate = "1:m"),
     "join keys did not fulfil 1:m validation"
   )
 
+  # lazy 1:m
   expect_error(
-    df2$join(df1, on = "x", validate = "foobar")$collect(),
+    df2$lazy()$join(df1$lazy(), on = "x", validate = "1:m")$collect(),
+    "join keys did not fulfil 1:m validation"
+  )
+
+  # eager error on unknown validate choice
+  expect_error(
+    df2$join(df1, on = "x", validate = "foobar"),
+    "should be one of"
+  )
+
+  # lazy error on unknown validate choice
+  expect_error(
+    df2$lazy()$join(df1$lazy(), on = "x", validate = "foobar")$collect(),
     "should be one of"
   )
 })
 
 test_that("argument 'join_nulls' works", {
-  df1 = pl$LazyFrame(x = c(NA, letters[1:2]), y = 1:3)
-  df2 = pl$LazyFrame(x = c(NA, letters[2:3]), y2 = 4:6)
+  df1 = pl$DataFrame(x = c(NA, letters[1:2]), y = 1:3)
+  df2 = pl$DataFrame(x = c(NA, letters[2:3]), y2 = 4:6)
 
-  # discard nulls by default
+  # 1 discard nulls by default
+
+  # eager1
   expect_identical(
-    df1$join(df2, on = "x")$collect()$to_data_frame(),
+    df1$join(df2, on = "x")$to_data_frame(),
     data.frame(x = "b", y = 3L, y2 = 5L)
   )
 
-  # consider nulls as a valid key
+  # lazy1
   expect_identical(
-    df1$join(df2, on = "x", join_nulls = TRUE)$collect()$to_data_frame(),
+    df1$lazy()$join(df2$lazy(), on = "x")$collect()$to_data_frame(),
+    data.frame(x = "b", y = 3L, y2 = 5L)
+  )
+
+  # 2 consider nulls as a valid key
+
+  # eager2
+  expect_identical(
+    df1$join(df2, on = "x", join_nulls = TRUE)$to_data_frame(),
     data.frame(x = c(NA, "b"), y = c(1L, 3L), y2 = c(4L, 5L))
   )
 
-  # several nulls
-  df2 = pl$LazyFrame(x = c(NA, letters[2:3], NA), y2 = 4:7)
+  # lazy2
   expect_identical(
-    df1$join(df2, on = "x", join_nulls = TRUE)$collect()$to_data_frame(),
+    df1$lazy()$join(df2$lazy(), on = "x", join_nulls = TRUE)$collect()$
+      to_data_frame(),
+    data.frame(x = c(NA, "b"), y = c(1L, 3L), y2 = c(4L, 5L))
+  )
+
+  # 3 several nulls
+  df3 = pl$DataFrame(x = c(NA, letters[2:3], NA), y2 = 4:7)
+
+  # eager3
+  expect_identical(
+    df1$join(df3, on = "x", join_nulls = TRUE)$to_data_frame(),
+    data.frame(x = c(NA, "b", NA), y = c(1L, 3L, 1L), y2 = c(4L, 5L, 7L))
+  )
+
+  # lazy3
+  expect_identical(
+    df1$lazy()$join(df3$lazy(), on = "x", join_nulls = TRUE)$collect()$
+      to_data_frame(),
     data.frame(x = c(NA, "b", NA), y = c(1L, 3L, 1L), y2 = c(4L, 5L, 7L))
   )
 })